feat(Refactoring): Refactor file structure
