+++ src/Components/Category/AllCategory.jsx
@@ -1,12 +1,12 @@
-import Croptop from './Croptop.jpg'
-import Fastrackwatch from './Fastrackwatch.webp'
-import Menshirt from './Menshirt.webp'
-import Menshoes from './Menshoes.jpg'
-import Suit from './Suit.jpg'
-import Suit2 from './Suit2.jpg'
-import Tshirt from './Tshirt.jpg'
-import Womenjeans from './Womenjeans.jpg'
-import WomenShoes from './WomenShoes.jpg'
+import Croptop from '../../assets/clothings/Croptop.jpg'
+import Fastrackwatch from '../../assets/clothings/Fastrackwatch.webp'
+import Menshirt from '../../assets/clothings/Menshirt.webp'
+import Menshoes from '../../assets/clothings/Menshoes.jpg'
+import Suit from '../../assets/clothings/Suit.jpg'
+import Suit2 from '../../assets/clothings/Suit2.jpg'
+import Tshirt from '../../assets/clothings/Tshirt.jpg'
+import Womenjeans from '../../assets/clothings/Womenjeans.jpg'
+import WomenShoes from '../../assets/clothings/WomenShoes.jpg'
 import ItemComponent from './ItemComponent'
 import {Link} from 'react-router-dom'
 const AllCategory = () => {
